[SVLS-9306] organize shared runtime copy
