+++ packages/base/src/helpers/serverless/ssi/constants.ts
@@ -1,0 +1,4 @@
+export const TRACER_COPY_CONTAINER_NAME = 'datadog-tracer-copy'
+export const TRACER_VOLUME_NAME = 'datadog-tracer'
+export const TRACER_MOUNT_PATH = '/datadog-lib'
+export const TRACER_READINESS_PORT = 18999
